7.18: kommentit, step1 ja 7.19: kommentit, step2

+++ Osa7/blogilista/frontend/src/services/blogs.js
@@ -20,11 +20,14 @@ const getAll = async () => {
 };
 
 const getById = async (id) => {
+  if (!id || id === 'undefined') {
+    throw new Error('invalid id');
+  }
   const response = await axios.get(`${baseUrl}/${id}`);
   const blog = response.data;
   return {
     ...blog,
-    id: blog._id,
+    id: blog.id ?? blog._id,
     user: blog.user
       ? { ...blog.user, id: blog.user._id }
       : null
@@ -40,10 +43,12 @@ const createBlog = async (newBlog) => {
 };
 
 const updateBlog = async (id, updatedBlog) => {
-  
   const config = {
     headers: { Authorization: token },
   };
+  if (!id || id === 'undefined') {
+    throw new Error('invalid id');
+  }
   const response = await axios.put(`${baseUrl}/${id}`, updatedBlog, config);
   const blog = response.data;
   return {
@@ -60,8 +65,27 @@ const deleteBlog = async (Id) => {
   const config = {
     headers: { Authorization: token },
   };
+  if (!Id || Id === 'undefined') {
+    throw new Error('invalid id');
+  }
   const response = await axios.delete(`${baseUrl}/${Id}`, config);
   return response.data;
 };
 
-export default { getAll, getById, setToken, createBlog, updateBlog, deleteBlog }
+//comments:
+const getComments = async (id) => {
+  if (!id || id === 'undefined') {
+    throw new Error('invalid id');
+  }
+  const response = await axios.get(`${baseUrl}/${id}/comments`);
+  return response.data;
+};
+const addComment = async (id, content) => {
+  if (!id || id === 'undefined') {
+    throw new Error('invalid id');
+  }
+  const response = await axios.post(`${baseUrl}/${id}/comments`, { content });
+  return response.data;
+};
+
+export default { getAll, getById, setToken, createBlog, updateBlog, deleteBlog, getComments, addComment }
